+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ## [5.2.3] - 2024-11-04
 
 - "Pacific Island" no longer matches Iceland.
+- "Island" no longer matches Iceland when language is not set.
 
 ## [5.2.2] - 2024-08-23